backend/requirements.txt
paho-mqtt>=1.6.1
pyttsx3>=2.90
